@@ -1,9 +1,9 @@
 import fse from 'fs-extra';
-import { tmpFolder } from './constants';
+import { tmpFolder } from './constants.js';
 
-export * from './utils';
+export * from './utils.js';
 // eslint-disable-next-line import/export
-export * from './constants';
+export * from './constants.js';
 
 export default class Test {
     async setTmpFolder() {
