@@ -1,5 +1,5 @@
 /* eslint-disable security/detect-non-literal-require */
-import { entry } from './constants';
+import { entry } from './constants.js';
 
 const m = require(entry);
 
