@@ -1,8 +1,14 @@
 import path from 'path';
 import { promisify } from 'util';
 import { exec } from 'child_process';
-import { entry } from './constants';
+import Module from 'module';
+import esm from 'esm';
+import { entry } from './constants.js';
 
+console.log('module:', typeof Module, typeof new Module());
+
+
+const require = esm(new Module());
 const execAsync = promisify(exec);
 
 export function load(relPath, clearCache) {
@@ -10,7 +16,10 @@ export function load(relPath, clearCache) {
 
     if (clearCache) delete require.cache[require.resolve(absPath)];
     // eslint-disable-next-line security/detect-non-literal-require
+    // console.log('require:', require);
     const result =  require(absPath);
+
+    console.log('result:', result);
 
     if (clearCache) delete require.cache[require.resolve(absPath)];
 
